Steganography/requirements.txt
Pillow==8.2.0
PyQT5==5.15.4